massage copy user

+++ manish/bigproject/php/massage/config.php
@@ -34,6 +34,12 @@ function saveData($firebase, $user_id, $data, $path1) {
     unset($data['paths']);
   }
   
+  $path = MAIN_PATH . '/users/'.$user_id;
+  $userData = json_decode($firebase->get($path), 1);
+  
+  $data['details']['postedBy'] = $userData['displayName'];
+  $data['details']['profileImage'] = $userData['image'];
+  
   $path = $path1 . '/records';
   $sid = $firebase->push($path, $data);
   $arr = json_decode($sid, 1);
@@ -53,10 +59,13 @@ function saveData($firebase, $user_id, $data, $path1) {
   $firebase->push($recordPath, 'location/' . base64_encode($data['location']['country']) . '/' . base64_encode($data['location']['state']) . '/' . base64_encode($data['location']['county']) . '/' . $pathID);
   
   
-  if (!empty($data['tags'])) {
-    $arr = explode(',', $data['tags']);
+  if (!empty($data['tags2'])) {
+    $arr = explode(',', $data['tags2']);
     foreach ($arr as $v) {
-      $v = strtolower(trim($v)); 
+      $v = strtolower(trim($v));
+      if (empty($v)) {
+       continue; 
+      }//end if
       if (!empty($data['location']['county'])) {
         $firebase->set($path1.'/tags/' . base64_encode($v) . '/' . base64_encode($data['location']['country']) . '/' . base64_encode($data['location']['state']) . '/' . base64_encode($data['location']['county']) . '/' . $pathID, time() * 1000);
         $firebase->push($recordPath, 'tags/' . base64_encode($v) . '/' .  base64_encode($data['location']['country']) . '/' . base64_encode($data['location']['state']) . '/' . base64_encode($data['location']['county']) . '/' . $pathID);
@@ -69,6 +78,10 @@ function saveData($firebase, $user_id, $data, $path1) {
 
 function deleteRecord($firebase, $record) {
   //delete all old paths
+  if (empty($record['paths'])) {
+    return;
+  }
+  
   foreach ($record['paths'] as $v) {
     $path = $record['sourcePath']. '/' . $v;
     $firebase->delete($path);
@@ -84,10 +97,12 @@ function deleteRecord($firebase, $record) {
 function copyRecordByUserId($firebase, $id, $user_id, $path1, $path2)
 {
   $record = getRecordDetails($firebase, $id, $path1, $path2);
+  if (empty($record)) {
+    return false; 
+  }
   //delete all old paths
   deleteRecord($firebase, $record);
   $record['uid'] = $user_id;
-  pr($record);
 
   //save data
   saveData($firebase, $user_id, $record, $path2);
